+++ R/data.R
@@ -1,0 +1,57 @@
+#' FDA Statin dataset with 46 adverse effects
+#'
+#' FAERS 2014Q3 - 2020Q4
+#'
+#' Includes 6 Statins:
+#'
+#' Atorvastatin, Fluvastatin, Lovastatin, Pravastatin, Rosuvastatin, Simvastatin
+#'
+#' Corresponding to 46 adverse effects (PTs) as given by FDA
+#'
+#' @source \url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
+"statin46"
+
+
+#' FDA Statin dataset with 1491 adverse effects
+#'
+#' FAERS 2014Q3 - 2020Q4
+#'
+#' Includes 6 Statins:
+#'
+#' Atorvastatin, Fluvastatin, Lovastatin, Pravastatin, Rosuvastatin, Simvastatin
+#'
+#' Corresponding to 1491 adverse effects (PTs) common between statins and GBCA
+#'
+#' @source \url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
+"statin1491"
+
+
+#' FDA Statin dataset with all observed 6039 adverse effects
+#'
+#' FAERS 2014Q3 - 2020Q4
+#'
+#' Includes 6 Statins:
+#'
+#' Atorvastatin, Fluvastatin, Lovastatin, Pravastatin, Rosuvastatin, Simvastatin
+#'
+#' Corresponding to all 6039 observed adverse effects (PTs) observed in statins
+#'
+#' @source \url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
+"statin"
+
+
+#' FDA GBCA dataset with all observed 1707 adverse effects
+#'
+#' FAERS 2014Q3 - 2020Q4
+#'
+#' Includes 6 GBCA's:
+#'
+#' gadobenate, gadobutrol, gadodiamide, gadofosveset, gadopentetate, gadoterate,
+#' gadoteridol, gadoversetamide, gadoxetate
+#'
+#' Corresponding to all 1707 observed adverse effects (PTs) observed in statins
+#'
+#' @source \url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
+"gbca"
+
+